# on 2004/05/31 09:58:12, mradwin did: Michael,
I seem to have found another error in your posting.  For 1st day
Shavuot,

4: Exodus 19:20-20:13 (8 p'sukim) <-- should be 19:20-20:14 (20 p'sukim)
5: Exodus 20:14-22 (9 p'sukim) <--should be 20:15-23

The verse numbering is according to the Kestenbaum tikkun and the USCJ
Luach and agrees with the aliyah breakup in the ArtScroll siddur.  Also,
I have seen that some sources number the verses differently through the
Ten Commandments (the "old" K'tav tikkun is an example).  The 4th aliyah
here goes through the tenth commandment.

Alan Ziv


git-svn-id: file:///home/hebcal/hebcal-dotcom-svn@1624 eef4b989-ff24-4261-96ef-5f3e0244f1be